Import libgtkhtml 2.6.2.
git-svn-id: svn+ssh://svn.mit.edu/athena/branches/vendor@20944 728af825-273c-0410-89f9-f7d3b574a069

+++ third/libgtkhtml/libgtkhtml/dom/html/dom-htmlinputelement.h
@@ -36,8 +36,8 @@
 
 struct _DomHTMLInputElement {
 	DomHTMLElement parent;
-	gint8 checked:2;
-	guint8 active:1;
+	gint checked:2;
+	guint active:1;
 	gchar *str_value;
 	/* FIXME: Not implemented:
 	  attribute DOMString        accept;
